#development-Pawel another try to push this shit

+++ TypicalMirek_UsedCarDealer/Logic/Managers/CarBodyManager.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-using TypicalMirek_UsedCarDealer.Logic.Factories;
+﻿using System.Linq;
 using TypicalMirek_UsedCarDealer.Logic.Factories.Interfaces;
 using TypicalMirek_UsedCarDealer.Logic.Managers.Interfaces;
 using TypicalMirek_UsedCarDealer.Logic.Repositories;
@@ -15,18 +11,26 @@ namespace TypicalMirek_UsedCarDealer.Logic.Managers
     {
         #region Properties
         private readonly IBodyRepository bodyRepository;
-
+        private readonly ICarRepository carRepository;
         #endregion
 
+        #region Constructors
         public CarBodyManager(IRepositoryFactory repositoryFactory)
         {
             bodyRepository = repositoryFactory.Get<BodyRepository>();
+            carRepository = repositoryFactory.Get<CarRepository>();
         }
+        #endregion
 
         public Body Add(Body body)
         {
             if (body.Id <= 0)
             {
+                if (bodyRepository.GetById(body.Id) != null || bodyRepository.CheckIfBodyWithExactNameExists(body.Name))
+                {
+                    return null;
+                }
+
                 bodyRepository.Add(body);
                 bodyRepository.Save();
             }
@@ -37,7 +41,13 @@ namespace TypicalMirek_UsedCarDealer.Logic.Managers
         public Body Modify(Body body)
         {
             var bodyToModify = bodyRepository.GetById(body.Id);
-            if (bodyToModify != null)
+            var isModyfiedNameEqual = body.Name.Equals(bodyToModify.Name);
+
+            if (bodyRepository.CheckIfBodyWithExactNameExists(body.Name) && !isModyfiedNameEqual)
+            {
+                return null;
+            }
+            else
             {
                 bodyToModify.Name = body.Name;
                 bodyRepository.Save();
@@ -46,14 +56,19 @@ namespace TypicalMirek_UsedCarDealer.Logic.Managers
             return bodyToModify;
         }
 
-        public void Delete(int id)
+        public bool Delete(int id)
         {
             var bodyToDelete = bodyRepository.GetById(id);
             if (bodyToDelete != null)
             {
-                bodyRepository.Delete(bodyToDelete);
-                bodyRepository.Save();
+                if (!carRepository.CheckIfExistCarForBodyId(id))
+                {
+                    bodyRepository.Delete(bodyToDelete);
+                    bodyRepository.Save();
+                    return true;
+                }
             }
+            return false;
         }
 
         public void Delete(Body body)
@@ -79,6 +94,7 @@ namespace TypicalMirek_UsedCarDealer.Logic.Managers
         public void Dispose()
         {
             bodyRepository.Dispose();
+            carRepository.Dispose();
         }
     }
 }
